# swagger.yaml
openapi: 3.0.3
info:
  title: Univote - University Voting System API
  description: |
    Complete API documentation for Univote, a secure university voting system with facial recognition, geofencing, and email notifications.
    
    ## Features
    - 🔐 JWT-based authentication for students and admins
    - 👤 Azure Face API facial verification
    - 📍 Geofencing validation (Bowen University campus)
    - 📧 Automated email notifications (welcome, vote confirmation, results)
    - 🗳️ Real-time voting with eligibility checks
    - 📊 Comprehensive analytics and reporting
    - 🔒 Single session enforcement
    - 📝 Audit logging for all critical operations
    
    ## Authentication Flow
    1. **First Login**: Use default password '1234' to get `first_login` token
    2. **Change Password**: Use first_login token to set new password (triggers welcome email)
    3. **Regular Login**: Login with new password to get `student` token
    4. **Access Protected Routes**: Use student/admin token in Authorization header
    
    ## Rate Limits
    - Authentication endpoints: 5 requests per 15 minutes
    - Voting endpoints: 10 requests per minute
    - Face API endpoints: 20 requests per minute
    - General API: 100 requests per 15 minutes
    
    ## Security
    - Passwords hashed with bcrypt (10 rounds)
    - JWT tokens expire after 24 hours
    - Single active session per user
    - Geofencing: 5000m radius from Bowen University (7.8525, 4.2811)
    
  version: 1.0.0
  contact:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from login endpoints. 
        Token types: 'student', 'admin', 'first_login'
  schemas:
...
    Student:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439011"
        matric_no:
          type: string
          example: "BU22CSC1005"
        email:
          type: string
          format: email
          example: "muhammedabiodun42@gmail.com"
        full_name:
          type: string
          example: "Mohammed Abiodun"
        college:
          type: string
          enum: [COAES, COMSS, COLAW, COLBS, COHES, COCCS, COEVS]
          example: "COCCS"
        department:
          type: string
          example: "Computer Science"
        level:
          type: string
          enum: ["100", "200", "300", "400", "500"]
          example: "200"
        first_login:
          type: boolean
          example: false
        face_person_id:
          type: string
          nullable: true
          example: "azure-face-api-person-id"
        session_token:
          type: string
          nullable: true
        last_login:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    Admin:
      type: object
      properties:
        _id:
          type: string
        email:
          type: string
          format: email
          example: "louisdiaz43@gmail.com"
        full_name:
          type: string
          example: "Louis Diaz"
        role:
          type: string
          enum: [admin, super_admin]
          example: "super_admin"
        session_token:
          type: string
          nullable: true
        last_login:
          type: string
          format: date-time
...
paths:
...
  /auth/admin-login:
    post:
      tags:
        - Authentication
      summary: Admin login
      description: Login as an administrator using email and password
      operationId: adminLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "louisdiaz43@gmail.com"
                password:
                  type: string
                  example: "balikiss12"
      responses:
        '200':
          description: Admin login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Admin login successful"
                  token:
                    type: string
                  token_type:
                    type: string
                    example: "admin"
                  admin:
                    $ref: '#/components/schemas/Admin'
        '401':
          description: Invalid credentials
        '429':
          description: Rate limit exceeded
  /auth/change-password:
    patch:
      tags:
        - Authentication
      summary: Change password
      description: |
        Change student password. Can be used with either:
        - `first_login` token (first-time password change)
        - `student` token (regular password change)
        
        **Important**: Welcome email is sent after first password change.
      operationId: changePassword
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - new_password
              properties:
                new_password:
                  type: string
                  minLength: 6
                  example: "mySecurePassword123"
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Password changed successfully"
                  token:
                    type: string
                    description: New student token (if first login)
                  token_type:
                    type: string
                    example: "student"
        '400':
          description: Validation error
        '401':
          description: Unauthorized
...
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get student profile
      description: Get current authenticated student's profile information
      operationId: getProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  student:
                    $ref: '#/components/schemas/Student'
        '401':
          description: Unauthorized

  # ==================== ADMIN - STUDENTS ====================
...
  /results/{session_id}:
    get:
      tags:
        - Results
      summary: Get session results
      description: |
        Get published results for a voting session.
        **Note**: Results are only visible after admin publishes them.
      operationId: getResults
      security:
        - BearerAuth: []
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
          description: Session ID
      responses:
        '200':
          description: Results retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  session:
                    type: object
                    properties:
                      title:
                        type: string
                      total_votes:
                        type: integer
                      published_at:
                        type: string
                        format: date-time
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        position:
                          type: string
                        total_votes:
                          type: integer
                        candidates:
                          type: array
                          items:
                            type: object
                            properties:
                              matric_no:
                                type: string
                              full_name:
                                type: string
                              votes:
                                type: integer
                              percentage:
                                type: number
                              is_winner:
                                type: boolean
        '401':
          description: Unauthorized
        '403':
          description: Results not yet published
        '404':
          description: Session not found
  /results/{session_id}/publish:
    post:
      tags:
        - Results
      summary: Publish results
      description: |
        Publish election results for a completed voting session.
        **Admin only**. Sends result announcement emails to all participants.
      operationId: publishResults
      security:
        - BearerAuth: []
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
          description: Session ID
      responses:
        '200':
          description: Results published successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  results:
                    type: object
        '400':
          description: Session not completed yet
        '401':
          description: Unauthorized (admin only)
        '404':
          description: Session not found
  /results/stats/overview:
    get:
      tags:
        - Results
      summary: Get overall statistics
      description: |
        Get comprehensive system-wide statistics. **Admin only**.
        Includes total students, sessions, votes, participation rates, and trends.
      operationId: getOverallStats
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Statistics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  stats:
                    type: object
                    properties:
                      total_students:
                        type: integer
                      total_sessions:
                        type: integer
                      total_votes:
                        type: integer
                      active_sessions:
                        type: integer
                      overall_participation_rate:
                        type: number
                      students_by_college:
                        type: object
                      students_by_level:
                        type: object
                      recent_activity:
                        type: array
        '401':
          description: Unauthorized (admin only)